Move `transform_response` to different file
