+++ crates/sncast/src/helpers/mod.rs
@@ -4,7 +4,6 @@ pub mod braavos;
 pub mod config;
 pub mod configuration;
 pub mod constants;
-pub mod data_transformer;
 pub mod fee;
 pub mod interactive;
 pub mod output_format;
