+++ crates/sncast/src/response/transformed_call.rs
@@ -1,11 +1,15 @@
 use super::command::CommandResponse;
+use crate::response::call::CallResponse;
 use crate::response::cast_message::SncastMessage;
+use anyhow::Result;
 use conversions::string::IntoHexStr;
+use data_transformer::reverse_transform_output;
 use foundry_ui::Message;
 use foundry_ui::styling;
 use serde::Serialize;
 use serde_json::Value;
 use serde_json::json;
+use starknet::core::types::{ContractClass, contract::AbiEntry};
 use starknet_types_core::felt::Felt;
 
 #[derive(Serialize, Clone)]
@@ -43,4 +47,32 @@ impl Message for SncastMessage<TransformedCallResponse> {
             })
         })
     }
+}
+
+#[must_use]
+pub fn transform_response(
+    result: &Result<CallResponse>,
+    contract_class: &ContractClass,
+    selector: &Felt,
+) -> Option<TransformedCallResponse> {
+    let Ok(CallResponse { response, .. }) = result else {
+        return None;
+    };
+
+    if response.is_empty() {
+        return None;
+    }
+
+    let ContractClass::Sierra(sierra_class) = contract_class else {
+        return None;
+    };
+
+    let abi: Vec<AbiEntry> = serde_json::from_str(sierra_class.abi.as_str()).ok()?;
+
+    let transformed_response = reverse_transform_output(response, &abi, selector).ok()?;
+
+    Some(TransformedCallResponse {
+        response_raw: response.clone(),
+        response: transformed_response,
+    })
 }
